fin_shop/pyproject.toml
[tool.poetry]
name = "fin_shop"
version = "0.1.0"
description = ""
authors = ["Vasker <v92prog@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "3.2.13"
django-modeltranslation = "^0.17.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
